+++ src/activities/reader/ReaderOptionsActivity.cpp
@@ -680,7 +680,7 @@ void ReaderOptionsActivity::buildOptionsScreen(UiApp::ScreenType& screen) {
   for (size_t i = 0; i < currentSettingsList.size(); ++i) {
     const auto& setting = currentSettingsList[i];
     if (settingShowsNavigationCaret(setting)) {
-      values[i] = ">";
+      values[i] = I18N.isRtl() ? "<" : ">";
     } else if (setting.type == SettingType::TOGGLE && setting.valuePtr != nullptr) {
       values[i] = SETTINGS.*(setting.valuePtr) ? tr(STR_STATE_ON) : tr(STR_STATE_OFF);
     } else if (setting.type == SettingType::ENUM && setting.valuePtr != nullptr) {